Move interaction focused gui tests to ui_tests
