+++ tests/integration_tests/share/test_ecl_run_new_config.py
