+++ tests/integration_tests/share/test_shell.py
