+++ tests/integration_tests/share/test_templating.py
