+++ tests/ui_tests/gui/baseline/test_histogram_log_scale-float.png
