+++ tests/ui_tests/gui/baseline/test_histogram_no_ensembles-empty.png
