+++ tests/ui_tests/gui/baseline/test_histogram_no_log_scale-empty.png
