+++ tests/ui_tests/gui/test_full_manual_update_workflow.py
