+++ tests/ui_tests/gui/test_main_window.py
