+++ tests/unit_tests/gui/simulation/view/test_realization.py
@@ -1,10 +1,41 @@
+from datetime import datetime as dt
+
+import pytest
 from qtpy import QtCore
 from qtpy.QtCore import QModelIndex, QSize
 from qtpy.QtWidgets import QStyledItemDelegate, QStyleOptionViewItem
 
+from ert.ensemble_evaluator.snapshot import (
+    EnsembleSnapshot,
+)
+from ert.ensemble_evaluator.state import (
+    FORWARD_MODEL_STATE_START,
+    REALIZATION_STATE_UNKNOWN,
+)
 from ert.gui.model.node import _Node
 from ert.gui.model.snapshot import SnapshotModel
 from ert.gui.simulation.view.realization import RealizationWidget
+from tests import SnapshotBuilder
+
+
+@pytest.fixture
+def small_snapshot() -> EnsembleSnapshot:
+    builder = SnapshotBuilder()
+    for i in range(0, 2):
+        builder.add_fm_step(
+            fm_step_id=str(i),
+            index=str(i),
+            name=f"job_{i}",
+            current_memory_usage="500",
+            max_memory_usage="1000",
+            status=FORWARD_MODEL_STATE_START,
+            stdout=f"job_{i}.stdout",
+            stderr=f"job_{i}.stderr",
+            start_time=dt(1999, 1, 1),
+            end_time=dt(2019, 1, 1),
+        )
+    real_ids = [str(i) for i in range(0, 5)]
+    return builder.build(real_ids, REALIZATION_STATE_UNKNOWN)
 
 
 class MockDelegate(QStyledItemDelegate):
